+++ crates/bankrun/Cargo.toml
@@ -19,6 +19,7 @@ solders-traits = { workspace = true, features = ["banks-client"]}
 solders-account = { workspace = true }
 solders-keypair = { workspace = true }
 solders-transaction = { workspace = true }
+solders-transaction-error = { workspace = true }
 solders-message = { workspace = true }
 solders-signature = { workspace = true }
 solders-transaction-status = { workspace = true }
