+++ crates/rpc-errors/Cargo.toml
@@ -12,6 +12,6 @@ serde = { workspace = true }
 derive_more = { workspace = true }
 solders-macros = { workspace = true }
 solders-traits = { workspace = true }
-solders-transaction-status = { workspace = true }
+solders-transaction-error = { workspace = true }
 solders-rpc-common = { workspace = true }
 solana-sdk = { workspace = true }
