+++ crates/rpc-errors/src/lib.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use solana_sdk::slot_history::Slot;
 use solders_macros::{common_methods, richcmp_eq_only, EnumIntoPy};
 use solders_traits::transaction_status_boilerplate;
-use solders_transaction_status::TransactionErrorType;
+use solders_transaction_error::TransactionErrorType;
 
 use solders_rpc_common::RpcSimulateTransactionResult;
