+++ crates/rpc-responses/Cargo.toml
@@ -18,6 +18,7 @@ solders-primitives = { workspace = true }
 solders-traits = { workspace = true }
 solders-account = { workspace = true }
 solders-account-decoder = { workspace = true }
+solders-transaction-error = { workspace = true }
 solders-transaction-status = { workspace = true }
 solders-rpc-common = { workspace = true }
 solders-rpc-errors = { workspace = true }
