+++ crates/transaction-status/Cargo.toml
@@ -23,5 +23,6 @@ solders-account-decoder = { workspace = true }
 solders-hash = { workspace = true }
 solders-message = { workspace = true }
 solders-transaction = { workspace = true }
+solders-transaction-error = { workspace = true }
 solders-commitment-config ={ workspace = true }
 bincode ={ workspace = true }
